Cleaned up backend code
