+++ go-backend/go.mod
@@ -2,4 +2,10 @@ module Identity-Lock/go-backend
 
 go 1.17
 
-require github.com/golang-jwt/jwt/v4 v4.1.0
+require (
+	github.com/golang-jwt/jwt/v4 v4.1.0
+	github.com/gorilla/handlers v1.5.1
+	github.com/gorilla/mux v1.8.0
+)
+
+require github.com/felixge/httpsnoop v1.0.1 // indirect
